Начисление, доработки по кол-ву проживающих

+++ src/main/java/com/dic/app/mm/ConfigApp.java
@@ -20,7 +20,7 @@ public interface ConfigApp {
 
 	Lock getLock();
 
-	boolean aquireLock(int rqn, int klskId);
+	boolean aquireLock(int rqn, long klskId);
 
 	int incNextReqNum();
